Lesson 26

+++ app/Http/Controllers/ThreadsController.php
@@ -114,6 +114,12 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+
+//        if ($thread->user_id != auth()->id()) {
+//            abort(403, 'You do not have the permission to do this.');
+//        }
+
         $thread->delete();
 
         if (request()->wantsJson()) {
